Refactor ATA sector reading and FAT handling; improve memory management
- Updated ReadSectors function in ata.c to use pointer arithmetic for writing data to TargetAddress.
- Simplified FATRead function in fat12.c to directly read from fatBuffer without casting.
- Added missing include for misc.h in kaosldr.c.
- Changed itoa_helper function signature in misc.c and misc.h to accept unsigned int instead of unsigned short.
- Added kernel.map generation to Rust Cargo configuration for better debugging.
- Adjusted Rust Cargo profile to set opt-level to 0 for improved debug information.
- Enhanced memory management in Rust kernel by implementing a new physical memory layout structure and functions for page frame allocation and release.
- Improved comments and documentation for clarity on memory layout and bitmap encoding.

+++ main64/kaosldr_64/misc.c
@@ -173,7 +173,7 @@ void itoa(int i, unsigned base, char *buf)
 
 // Helper function for the itoa function.
 // The static keyword means that this function is only available within the scope of this object file.
-static void itoa_helper(unsigned short i, unsigned base, char *buf)
+static void itoa_helper(unsigned int i, unsigned base, char *buf)
 {
     int pos = 0;
     int opos = 0;
